generalize ConvexSet methods to LazySet
