+++ src/Initialization/init_Polyhedra.jl
@@ -49,3 +49,4 @@ eval(load_polyhedra_hpolyhedron())
 eval(load_polyhedra_vpolytope())
 eval(load_polyhedra_universe())
 eval(load_polyhedra_mesh())
+eval(load_polyhedra_lazyset())
